pyproject.toml
[project]
name = "modules"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi-users[jwt,sqlalchemy]>=15.0.3",
    "librosa==0.11.0",
    "mutagen>=1.47.0",
    "nemo-toolkit[asr]==2.6.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[bcrypt]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.2.1",
    "soundfile==0.13.1",
    "tenacity>=9.1.4",
    "torch==2.6.0",
    "uvicorn>=0.40.0",
    "yt-dlp>=2025.12.8",
]
